move metrics query to where it's used
