+++ src/Beckett.Dashboard/Home/Index.razor
@@ -1,5 +1,8 @@
+@using Beckett.Dashboard.Metrics
+@using Beckett.Database
 @layout Layout
-@inject IDashboard DashboardService
+@inject IPostgresDatabase Database
+@inject PostgresOptions PostgresOptions
 
 <div id="metrics" class="row mt-4 mx-2" hx-get="@Dashboard.Prefix/" hx-trigger="every 10s" hx-select="#metrics"
      hx-swap="outerHTML">
@@ -43,7 +46,7 @@
 
   protected override async Task OnInitializedAsync()
   {
-    var result = await DashboardService.Metrics.GetSubscriptionMetrics(CancellationToken.None);
+    var result = await Database.Execute(new MetricsQuery(PostgresOptions), CancellationToken.None);
 
     Model = new ViewModel(result.Lagging, result.Retries, result.Failed);
   }
