--- src/Beckett.Dashboard/IDashboardMetrics.cs
@@ -1,8 +1,0 @@
-namespace Beckett.Dashboard;
-
-public interface IDashboardMetrics
-{
-    Task<GetSubscriptionMetricsResult> GetSubscriptionMetrics(CancellationToken cancellationToken);
-}
-
-public record GetSubscriptionMetricsResult(long Lagging, long Retries, long Failed);
