+++ src/Beckett.Dashboard/Metrics/Metrics.razor
@@ -1,5 +1,7 @@
+@using Beckett.Database
 @inherits Component<Metrics.ViewModel>
-@inject IDashboard DashboardService
+@inject IPostgresDatabase Database
+@inject PostgresOptions PostgresOptions
 
 <a href="@Dashboard.Prefix/subscriptions/checkpoints/lagging"
    class="badge text-bg-info text-decoration-none"
@@ -28,7 +30,7 @@
       return;
     }
 
-    var result = await DashboardService.Metrics.GetSubscriptionMetrics(CancellationToken.None);
+    var result = await Database.Execute(new MetricsQuery(PostgresOptions), CancellationToken.None);
 
     Model = new ViewModel(result.Lagging, result.Retries, result.Failed, false);
   }
