+++ src/Beckett.Dashboard/Metrics/MetricsEndpoint.cs
@@ -1,10 +1,16 @@
+using Beckett.Database;
+
 namespace Beckett.Dashboard.Metrics;
 
 public static class MetricsEndpoint
 {
-    public static async Task<IResult> Handle(IDashboard dashboard, CancellationToken cancellationToken)
+    public static async Task<IResult> Handle(
+        IPostgresDatabase database,
+        PostgresOptions options,
+        CancellationToken cancellationToken
+    )
     {
-        var result = await dashboard.Metrics.GetSubscriptionMetrics(cancellationToken);
+        var result = await database.Execute(new MetricsQuery(options), cancellationToken);
 
         return Results.Extensions.Render<Metrics>(
             new Metrics.ViewModel(result.Lagging, result.Retries, result.Failed, false)
