+++ src/Beckett.Dashboard/ServiceCollectionExtensions.cs
@@ -1,4 +1,3 @@
-using Beckett.Dashboard.Postgres.Metrics;
 using Beckett.Dashboard.Postgres.Subscriptions;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -8,7 +7,6 @@ public static class ServiceCollectionExtensions
 {
     internal static void AddDashboardSupport(this IServiceCollection services)
     {
-        services.AddSingleton<IDashboardMetrics, PostgresDashboardMetrics>();
         services.AddSingleton<IDashboardSubscriptions, PostgresDashboardSubscriptions>();
         services.AddSingleton<IDashboard, DefaultDashboard>();
     }
